acertpix-api-facematch/pyproject.toml
[project]
name = "acertpix_api_facematch"
version = "0.1.0"
description = "Servidor MCP para integração com API FaceMatch da Acertpix"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "mcp>=1.6.0",
    "httpx>=0.28.1",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
]
[[project.authors]]
name = "Marcelo Cabral Ghilardi"
email = "marcelo.cabral@acertpix.com.br"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
acertpix-api-facematch = "acertpix_api_facematch:main"

[tool.hatch.build.targets.wheel]
packages = ["src/acertpix_api_facematch"]
